requirements.txt
pyTelegramBotAPI==4.16.1
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
